python/requirements.txt
confluent-kafka==1.7.0
msal==1.12.0
python-dotenv==0.17.1
